web_ui_utcp_mcp_bridge/requirements.txt
utcp==0.2.1
fastmcp
fastapi
python-dotenv